create base EventStreamHandler class
will make things easier when we have two event stream emitters
